Überprüfung ob Restaurant

+++ hm.edu.webcrawler/src/main/RestaurantCrawler.py
@@ -75,6 +75,7 @@ def navigateToRestaurantDetailPage(restaurantName , city):
             googleElement.click()
     finally:        
         
+        
         try:  
             present = False
             try:
@@ -103,37 +104,46 @@ def navigateToRestaurantDetailPage(restaurantName , city):
         
 def crawlData(driver, wait):     
     
-    restaurant_title = wait.until(EC.element_to_be_clickable((By.CLASS_NAME, 'section-hero-header-title'))).get_attribute("innerHTML")
-    restaurant_stars = wait.until(EC.element_to_be_clickable((By.CLASS_NAME, 'section-star-display'))).get_attribute("innerHTML")
-    restaurant_rushHour = ''
-    numberOfReviewsButton = driver.find_element_by_class_name('section-reviewchart-numreviews')
-    numberOfReviews = numberOfReviewsButton.get_attribute("innerHTML")[0:-9]
-    if('.' in numberOfReviews):
-        numberOfReviews = int(numberOfReviews.replace('.', ''))
-    else:
-        numberOfReviews = int(numberOfReviews)
+    #googlePlace = wait.until(EC.element_to_be_clickable((By.CLASS_NAME, 'section-rating-term'))).find_element_by_class_name('widget-pane-link').get_attribute("innerHTML")
+    googlePlace = wait.until(EC.element_to_be_clickable((By.XPATH, '//*[@id="pane"]/div/div[1]/div/div/div/div[1]/div[3]/div[2]/div/div[2]/span[1]/span[1]/button'))).get_attribute("innerHTML")
+    #googlePlace = wait.until(EC.element_to_be_clickable((By.CLASS_NAME, 'widget-pane-link'))).get_attribute("innerHTML")
+    
+    if('Restaurant' in googlePlace):
         
+        restaurant_title = wait.until(EC.element_to_be_clickable((By.CLASS_NAME, 'section-hero-header-title'))).get_attribute("innerHTML")
+        restaurant_stars = wait.until(EC.element_to_be_clickable((By.CLASS_NAME, 'section-star-display'))).get_attribute("innerHTML")
+        restaurant_rushHour = ''
+        numberOfReviewsButton = driver.find_element_by_class_name('section-reviewchart-numreviews')
+        numberOfReviews = numberOfReviewsButton.get_attribute("innerHTML")[0:-9]
+        if('.' in numberOfReviews):
+            numberOfReviews = int(numberOfReviews.replace('.', ''))
+        else:
+            numberOfReviews = int(numberOfReviews)
+            
+            
+        print(restaurant_title)
+        print(restaurant_stars)
+        print(numberOfReviews)
+            
+        numberOfReviewsButton.click()
+        reviewDetailPage = wait.until(EC.element_to_be_clickable((By.CLASS_NAME, 'section-header-title')))
+        if reviewDetailPage.get_attribute("innerHTML") == 'Alle Rezensionen':
+            # Crawl Comments
+            
+            db = TinyDB('..\database\db.json')
+            db.purge()
+            restaurantTable = db.table('RESTAURANT TABLE')
+            reviewTable = db.table('REVIEW TABLE')
+            reviewPicturesTable = db.table('REVIEW PICTURE TABLE')
+            
+            reviewList = scrollOverAllReviews(driver, 0.1, wait, numberOfReviews)
+            
+            restaurant = Restaurant(restaurant_title, restaurant_stars, restaurant_rushHour , reviewList)
+            
+            insertReviewIntoDB(restaurant, restaurantTable, reviewTable, reviewPicturesTable)   
+    else: 
+        print('Gefundener Ort ist kein Restaurant')
         
-    print(restaurant_title)
-    print(restaurant_stars)
-    print(numberOfReviews)
-        
-    numberOfReviewsButton.click()
-    reviewDetailPage = wait.until(EC.element_to_be_clickable((By.CLASS_NAME, 'section-header-title')))
-    if reviewDetailPage.get_attribute("innerHTML") == 'Alle Rezensionen':
-        # Crawl Comments
-        
-        db = TinyDB('..\database\db.json')
-        db.purge()
-        restaurantTable = db.table('RESTAURANT TABLE')
-        reviewTable = db.table('REVIEW TABLE')
-        reviewPicturesTable = db.table('REVIEW PICTURE TABLE')
-        
-        reviewList = scrollOverAllReviews(driver, 0.1, wait, numberOfReviews)
-        
-        restaurant = Restaurant(restaurant_title, restaurant_stars, restaurant_rushHour , reviewList)
-        
-        insertReviewIntoDB(restaurant, restaurantTable, reviewTable, reviewPicturesTable)    
     
         
 def insertReviewIntoDB(restaurant, restaurantTable, reviewTable, reviewPicturesTable):
